fix(some-serial): bound PL011 RX discard
