--- drivers/serial/some-serial/.qemu.toml
@@ -1,11 +1,0 @@
-args = [
-  "-nographic",
-  "-cpu",
-  "cortex-a53",
-  "-machine",
-  "virt,gic-version=3,virtualization=on",
-]
-uefi = false
-to_bin = true
-success_regex = ["All tests passed"]
-fail_regex = ["Panicked:", "thread 'main' has overflowed its stack"]
