+++ drivers/serial/some-serial/Cargo.toml
@@ -4,7 +4,6 @@ categories = ["embedded", "hardware-support", "no-std"]
 description = "Unified serial driver collection for embedded and bare-metal environments"
 documentation = "https://docs.rs/some-serial"
 edition.workspace = true
-autotests = false
 keywords = ["serial", "uart", "embedded", "bare-metal"]
 license = "MIT OR Apache-2.0"
 name = "some-serial"
@@ -22,11 +21,6 @@ tock-registers.workspace = true
 
 [target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dependencies]
 x86.workspace = true
-
-# bare-test integration test is temporarily disabled.
-# [[test]]
-# harness = false
-# name = "test"
 
 [package.metadata.docs.rs]
 all-features = true
