--- drivers/serial/some-serial/test_build.sh
@@ -1,2 +1,0 @@
-cargo build
-cargo build --target x86_64-unknown-none
